pyproject.toml
[project]
name = "mdvupy"
version = "0.1.0"
description = "Cross-platform Markdown viewer in Python + Qt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"pyside6>=6.7",
	"markdown-it-py>=3.0.0",
	"pyqtdarktheme>=2.1.0",
	"loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
	"pytest",
	"pytest-qt",
	"ruff",
	"mkdocs",
]

[build-system]
requires = ["hatchling>=1.22"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mdvupy"]
